public version 1.0.0

+++ server/src/main/java/com/github/yeecode/matrixauth/server/business/AuthBusiness.java
@@ -4,13 +4,11 @@ import com.github.yeecode.matrixauth.server.cacheclient.CacheClient;
 import com.github.yeecode.matrixauth.server.constant.RequestSource;
 import com.github.yeecode.matrixauth.server.constant.RoleType;
 import com.github.yeecode.matrixauth.server.constant.Sentence;
-import com.github.yeecode.matrixauth.server.dao.ApplicationDao;
-import com.github.yeecode.matrixauth.server.dao.AuthDao;
-import com.github.yeecode.matrixauth.server.dao.RoleDao;
-import com.github.yeecode.matrixauth.server.dao.RoleXPermissionDao;
-import com.github.yeecode.matrixauth.server.dao.UserXRoleDao;
+import com.github.yeecode.matrixauth.server.dao.*;
 import com.github.yeecode.matrixauth.server.model.ApplicationModel;
+import com.github.yeecode.matrixauth.server.model.PermissionModel;
 import com.github.yeecode.matrixauth.server.model.RoleModel;
+import com.github.yeecode.matrixauth.server.model.UserModel;
 import com.github.yeecode.matrixauth.server.tenant.TenantSwitcher;
 import com.github.yeecode.matrixauth.server.util.FullUserKeyGenerator;
 import com.github.yeecode.matrixauth.server.util.Result;
@@ -24,10 +22,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 @Service
 public class AuthBusiness {
+    @Autowired
+    private UserDao userDao;
+    @Autowired
+    private PermissionDao permissionDao;
     @Autowired
     private RoleDao roleDao;
     @Autowired
@@ -60,8 +63,9 @@ public class AuthBusiness {
                 return ResultUtil.getFailResult(Sentence.NONE_APP_FOUND);
             }
             CacheClient cacheClient = tenantSwitcher.switchByApplication(applicationModel);
+            UserModel userModel = userDao.queryByKey(userKey,appName);
             RoleModel roleModel = roleDao.queryByName(roleName,appName);
-            if (roleModel == null) {
+            if (userModel == null || roleModel == null) {
                 return ResultUtil.getFailResult(Sentence.NONE_OBJECT_FOUND);
             }
 
@@ -88,7 +92,12 @@ public class AuthBusiness {
                 return ResultUtil.getFailResult(Sentence.NONE_APP_FOUND);
             }
             CacheClient cacheClient = tenantSwitcher.switchByApplication(applicationModel);
-            RoleModel roleModel = roleDao.queryByName(appName, roleName);
+            UserModel userModel = userDao.queryByKey(userKey,appName);
+            RoleModel roleModel = roleDao.queryByName(roleName,appName);
+            if (userModel == null || roleModel == null) {
+                return ResultUtil.getFailResult(Sentence.NONE_OBJECT_FOUND);
+            }
+
             if (RoleType.valueOf(roleModel.getType()).isSupportedRequestSources(requestSource)) {
                 if (tokenValidator.checkApplicationToken(appToken, applicationModel)) {
                     Integer count = proxySelf.deleteUserXRole(appName, userKey, roleName, cacheClient);
@@ -115,7 +124,12 @@ public class AuthBusiness {
 
             if (tokenValidator.checkApplicationToken(appToken, applicationModel)) {
                 CacheClient cacheClient = tenantSwitcher.switchByApplication(applicationModel);
-                Integer count = proxySelf.addRoleXPermission(appToken, roleName, permKey, cacheClient);
+                RoleModel roleModel = roleDao.queryByName(roleName,appName);
+                PermissionModel permissionModel = permissionDao.queryByKey(permKey,appName);
+                if (roleModel == null || permissionModel == null) {
+                    return ResultUtil.getFailResult(Sentence.NONE_OBJECT_FOUND);
+                }
+                Integer count = proxySelf.addRoleXPermission(appName, roleName, permKey, cacheClient);
                 return ResultUtil.getSuccessResult(count);
             } else {
                 return ResultUtil.getFailResult(Sentence.ILLEGAL_APP_TOKEN);
@@ -126,7 +140,7 @@ public class AuthBusiness {
         }
     }
 
-    public Result delRoleXPermission(String appToken, String appName, String roleName, String permKey) {
+    public Result deleteRoleXPermission(String appToken, String appName, String roleName, String permKey) {
         try {
             ApplicationModel applicationModel = applicationDao.queryByName(appName);
             if (applicationModel == null) {
@@ -135,7 +149,12 @@ public class AuthBusiness {
 
             if (tokenValidator.checkApplicationToken(appToken, applicationModel)) {
                 CacheClient cacheClient = tenantSwitcher.switchByApplication(applicationModel);
-                Integer count = proxySelf.delRoleXPermission(appName, roleName, permKey, cacheClient);
+                RoleModel roleModel = roleDao.queryByName(roleName,appName);
+                PermissionModel permissionModel = permissionDao.queryByKey(permKey,appName);
+                if (roleModel == null || permissionModel == null) {
+                    return ResultUtil.getFailResult(Sentence.NONE_OBJECT_FOUND);
+                }
+                Integer count = proxySelf.deleteRoleXPermission(appName, roleName, permKey, cacheClient);
                 return ResultUtil.getSuccessResult(count);
             } else {
                 return ResultUtil.getFailResult(Sentence.ILLEGAL_APP_TOKEN);
@@ -189,7 +208,7 @@ public class AuthBusiness {
     }
 
     @Transactional
-    public Integer delRoleXPermission(String appName, String roleName, String permKey, CacheClient cacheClient) {
+    public Integer deleteRoleXPermission(String appName, String roleName, String permKey, CacheClient cacheClient) {
         List<String> affectedUserKeys = authDao.queryUserKeyByRoleName(roleName, appName);
         Integer count = roleXPermissionDao.delete(appName, roleName, permKey);
         refreshUserXPermissions(appName, affectedUserKeys, cacheClient);
@@ -198,7 +217,7 @@ public class AuthBusiness {
 
     private void refreshUserXPermissions(String appName, List<String> affectedUserKeys, CacheClient cacheClient) {
         for (String userKey : affectedUserKeys) {
-            String permissionKeys = StringUtils.join(authDao.queryPermKeysByUserKey(appName,userKey), ';');
+            String permissionKeys = StringUtils.join(new HashSet<>(authDao.queryPermKeysByUserKey(appName,userKey)), ';');
             String fullUserKey = FullUserKeyGenerator.getFullUserKey(appName,userKey);
             userXPermissionBusiness.addOrUpdatePermissionsByFullUserKey(fullUserKey, permissionKeys);
             cacheClient.addOrUpdate(fullUserKey, permissionKeys);
